Conplete.jl Documentation

Conplete.transform — Method

transform(instance, target_type) -> target_instance

Transform instance of an NP complete problem into target_instance of target_type type.

It is only possible if there exists a chain of transformations from instance type into target_type type which is recorded in the package's transformation graph.

When transformation of solutions is needed chain_transform should be used instead.

Check add_problem and add_transformation for adding transformations into transformation graph

Arguments

instance::NPProblem: input instance of the problem.

target_type::Type{<:NPProblem}: target type of the output instance of the problem.

Examples

julia> using Conplete
julia> transform(SAT3([1 2 3]), VertexCover)
Instance of problem VertexCover

Conplete.transform — Method

transform(instance , chain_path) -> target_instance

Transform instance of an NP complete according to chain_path, while returning only the last instance.

It is only possible if there exists a chain of transformations from instance type matching chain_path which is recorded in the package's transformation graph.

When transformation of solutions is needed chain_transform should be used instead.

Check add_problem and add_transformation for adding transformations into transformation graph

Arguments

instance::NPProblem: input instance of the problem.

chain_path::Array{DataType}: chain of conversions as an array of problem types.

Examples

```jldoctest julia> using Conplete julia> transform(SAT3([1 2 3]), [SubsetSum,Partition,Knapsack]) Instance of problem Knapsack

Conplete.chain_transform — Method

chain_transform(instance, target_type) -> chain

Transform instance of an NP complete problem into instance of target_type type, while returning whole chain of intermediate problems.

It is only possible if there exists a chain of transformations from instance type into target_type type which is recorded in the package's transformation graph.

Check add_problem and add_transformation for adding transformations into transformation graph

Arguments

instance::NPProblem: input instance of the problem.

target_type::Type{<:NPProblem}: target type of the last transformation

Examples

jldoctest julia> using Conplete julia> chain_transform(SAT3([1 2 3]), Knapsack) 4-element Vector{NPProblem}: Instance of problem SAT3 Instance of problem SubsetSum Instance of problem Partition Instance of problem Knapsack

Conplete.chain_transform — Method

chain_transform(instance , chain_path) -> chain

Transform instance of an NP complete according to chain_path, while returning whole chain of intermediate problems.

It is only possible if there exists a chain of transformations from instance type matching chain_path which is recorded in the package's transformation graph.

Check add_problem and add_transformation for adding transformations into transformation graph

Arguments

instance::NPProblem: input instance of the problem.

chain_path::Array{DataType}: chain of conversions as an array of problem types.

Examples

julia> using Conplete
julia> chain_transform(SAT3([1 2 3]), [SubsetSum,Partition,Knapsack])
4-element Vector{NPProblem}:
 Instance of problem SAT3
 Instance of problem SubsetSum
 Instance of problem Partition
 Instance of problem Knapsack

Conplete.extract — Method

extract(solution, instance_chain) -> target_solution

From solution of the child problem extract the solution of the parent problem that was base for child problem when transformation was performed.

instance_chain contains all problem instances that were produced during chain transformation

Depth specifies how deep to extract when problem is a result of multiple transformations.

Conplete.construct — Method

construct(solution, chain) -> target_solution

Using solution of the base problem, construct the target_solution for problem that is result of the transformation of said base problem.

chain includes base problem, target problem and all inbetween problem instances computed during transformation, all of them are needed for this operation.

Conplete.construct — Method

construct(target_type, solution, problem) -> target_solution

Using solution of the base problem, construct the target_solution for problem of target_type that is result of the transformation of said base problem.

If type of solution and target_type are separated in transformation graph by more than one transformation, the instance problem needs to be transformed which will be slower

Therefore it is advised to use the method construct(solution::NPSolution, chain::Array{NPProblem}) in this situation.

See also: chain_transform

Conplete.solve — Function

Solve problem with JuMP using given solver

Conplete.add_problem — Method

add_problem(instance_type, solution_type)

Add a custom NP complete problem to the transformation graph. This makes it available for using add_transformation.

Conplete.add_transformation — Method

add_transformation(new, parent)

Add a transformation to the transformation graph, This makes it available for transformation using transform and chain_transform functions.