Library interfaces

Library utilizes two sets of interfaces for two distinct use cases.

For simple reductions regular interfaces can be used

Conplete.transform — Method

transform(instance, target_type) -> target_instance

Transform instance of an NP complete problem into target_instance of target_type type.

It is only possible if there exists a chain of transformations from instance type into target_type type which is recorded in the package's transformation graph.

When transformation of solutions is needed chain_transform should be used instead.

Check add_problem and add_transformation for adding transformations into transformation graph

Arguments

instance::NPProblem: input instance of the problem.

target_type::Type{<:NPProblem}: target type of the output instance of the problem.

Examples

julia> using Conplete
julia> transform(SAT3([1 2 3]), VertexCover)
Instance of problem VertexCover

Conplete.transform — Method

transform(instance , chain_path) -> target_instance

Transform instance of an NP complete according to chain_path, while returning only the last instance.

It is only possible if there exists a chain of transformations from instance type matching chain_path which is recorded in the package's transformation graph.

When transformation of solutions is needed chain_transform should be used instead.

Check add_problem and add_transformation for adding transformations into transformation graph

Arguments

instance::NPProblem: input instance of the problem.

chain_path::Vector{DataType}: chain of conversions as an Vector of problem types.

Examples

julia> using Conplete
julia> transform(SAT3([1 2 3]), [SubsetSum,Partition,Knapsack])
Instance of problem Knapsack

Conplete.extract — Function

extract(solution, parent_problem) -> parent_solution

From solution of the problem instance that was result of the transformation, extract solution of the parent problem.

See also: chain_transform

Examples

julia> using Conplete
julia> extract(VertexCoverSolution(Set([1,3,6])), SAT3([1 2 3]))
SAT3Solution(Bool[1, 0, 1])

Conplete.construct — Function

construct(target_type, solution, problem) -> target_solution

Using solution of the base problem, construct the target_solution for problem of target_type that is result of the transformation of said base problem.

See also: chain_transform

Examples

julia> using Conplete
julia> construct(VertexCover, SAT3Solution(Bool[1,1,0]), SAT3([1 2 3]))
VertexCoverSolution{Set{Int64}}(Set([6, 2, 9, 8, 1]))

Conplete.solve — Function

solve(solver, problem) -> solution

Solve problem with JuMP using given solver JuMP is required to be added for methos to appear

Examples

julia> using Conplete, JuMP, HiGHS
julia> solve(HiGHS.Optimizer, SAT3([1 2 3]))
SAT3Solution(Bool[1, 1, 0])

Chain reductions

It is possible to perform reductions that chain several reduction algorithms, when performing such reductions you should use interfaces described below.

When performing problem instance transformation you can also use regular transform function, but for solution transformations chain functions are mandatory

Conplete.chain_transform — Method

chain_transform(instance, target_type) -> chain

Transform instance of an NP complete problem into instance of target_type type, while returning whole chain of intermediate problems.

It is only possible if there exists a chain of transformations from instance type into target_type type which is recorded in the package's transformation graph.

Check add_problem and add_transformation for adding transformations into transformation graph

Arguments

instance::NPProblem: input instance of the problem.

target_type::Type{<:NPProblem}: target type of the last transformation

Examples

julia> using Conplete
julia> chain_transform(SAT3([1 2 3]), Knapsack)
4-element Vector{NPProblem}:
 Instance of problem SAT3
 Instance of problem SubsetSum
 Instance of problem Partition
 Instance of problem Knapsack

Conplete.chain_transform — Method

chain_transform(instance , chain_path) -> chain

Transform instance of an NP complete according to chain_path, while returning whole chain of intermediate problems.

It is only possible if there exists a chain of transformations from instance type matching chain_path which is recorded in the package's transformation graph.

Check add_problem and add_transformation for adding transformations into transformation graph

Arguments

instance::NPProblem: input instance of the problem.

chain_path::Vector{DataType}: chain of conversions as an Vector of problem types.

Examples

julia> using Conplete
julia> chain_transform(SAT3([1 2 3]), [SubsetSum,Partition,Knapsack])
4-element Vector{NPProblem}:
 Instance of problem SAT3
 Instance of problem SubsetSum
 Instance of problem Partition
 Instance of problem Knapsack

Conplete.extract — Method

extract(solution, instance_chain) -> target_solution

From solution of the child problem extract the solution of the parent problem that was base for child problem when transformation was performed.

instance_chain contains all problem instances that were produced during chain transformation

Depth specifies how deep to extract when problem is a result of multiple transformations.

Examples

julia> using Conplete
julia> x = chain_transform(SAT3([1 2 3]), HittingSet)
3-element Vector{NPProblem}:
 Instance of problem SAT3
 Instance of problem VertexCover
 Instance of problem HittingSet{Set{Int64}}
julia> extract(HittingSetSolution(Set([1,3,6])), x)
SAT3Solution(Bool[1, 0, 1])

Conplete.construct — Method

construct(solution, chain) -> target_solution

Using solution of the base problem, construct the target_solution for problem that is result of the transformation of said base problem.

chain includes base problem, target problem and all inbetween problem instances computed during transformation, all of them are needed for this operation.

Examples

julia> using Conplete
julia> x = chain_transform(SAT3([1 2 3]), HittingSet)
3-element Vector{NPProblem}:
 Instance of problem SAT3
 Instance of problem VertexCover
 Instance of problem HittingSet{Set{Int64}}
julia> construct(SAT3Solution(Bool[1,1,0]), x)
HittingSetSolution{Set{Int64}}(Set([6, 2, 9, 8, 1]))