Problem structures

Conplete.DirHamCycleSolution — Type

Solution to a Hamiltonian-Cycle problem containing cycle in following format:

an Vector of length equal to number of vertices where value cycle[x] = y corresponds to edge (x,y) in cycle

Conplete.HamCycleSolution — Type

Solution to a Hamiltonian-Cycle problem containing cycle in following format:

an Vector of length equal to number of vertices where value cycle[x] = y corresponds to edge (x,y) in cycle

Conplete.SubsetSumSolution — Type

Solution to a Subset Sum problem containing a set of indices of selected elements

Conplete.TSPSolution — Type

Solution to a TravellingSalesman problem containing cycle in following format:

an Vector of length equal to number of vertices where value cycle[x] = y corresponds to edge (x,y) in cycle