Problem structures

Conplete.BinPacking — Type

Instance of a Bin Packing problem

It utilizes generic integers, and can be constructed with any type which is a child of Integer abstract type

Fields

elements : vector of integers of which we seek a packing

bins : number of available bins

bin_size : size of a single bin

Conplete.BinPackingSolution — Type

Solution to a BinPacking problem

assigment[x] = y denotes assigment of element x to a bin y

Fields

assigment : vector denoting assigment of a element to a bin

Conplete.CNFSAT — Type

Instance of a Satisfiability problem in Conjunctive normal form.

Variables are denoted by integers from 1 to variable_count, negations of variables are denoted by corresponding negative integers.

Fields

variable_count : Number of different variables present in clauses

clauses : clauses denoted as a vector of clause vectors

Conplete.CNFSATSolution — Type

Solution to a CNFSAT problem

Fields

evaluation : BitVector denoting evaluation of each variable

Conplete.Clique — Type

Instance of a Clique problem.

Fields

graph : Undirected graph in which we seek a clique

size : minimal size of a clique

Conplete.CliqueSolution — Type

Solution to a Clique problem

Fields

clique : A set containing integers, corresponding to vertices belonging to clique

Conplete.DirHamCycle — Type

Instance of a Hamiltonian Cycle problem in a directed graph.

Fields

graph : Directed graph in which we seek a Hamiltonian Cycle

Conplete.DirHamCycleSolution — Type

Solution to a DirHamCycle problem

Fields

cycle : Vector of length equal to number of vertices where value cycle[x] = y corresponds to edge (x,y) in cycle

Conplete.HamCycle — Type

Instance of a Hamiltonian Cycle problem in a undirected graph.

Fields

graph : Undirected graph in which we seek a Hamiltonian Cycle

Conplete.HamCycleSolution — Type

Solution to a HamCycle problem

Fields

cycle : Vector of length equal to number of vertices where value cycle[x] = y corresponds to edge (x,y) in cycle

Conplete.HittingSet — Type

Instance of a Hitting Set problem

Set elements are denoted by the integers from 1 to universe_size

Fields

universe_size : number of different elements that can be present in sets

sets : vector of sets, in which we seek a Hitting Set

size : maximal size of a hitting set

Conplete.HittingSetSolution — Type

Solution to a HittingSet problem

Fields

subset : generic set of elements contained in hitting set

Conplete.Knapsack — Type

Instance of a Knapsack problem

It utilizes generic integers, and can be constructed with any type which is a child of Integer abstract type

Fields

elements : vector of integers denoting element size

elements : vector of integers denoting element value

size : size of a knapsack (maximal sum of elements)

min_value : minimal value sum of elements in knapsack

Conplete.KnapsackSolution — Type

Solution to a Knapsack problem

Fields

subset : generic set of indices corresponding to elements that are included in Knapsack

Conplete.Partition — Type

Instance of a Partition problem

It utilizes generic integers, and can be constructed with any type which is a child of Integer abstract type

Fields

elements : vector of integers of which we seek a partition

Conplete.PartitionSolution — Type

Solution to a Partition problem

Fields

subset : generic set of indices corresponding to elements that are present in one part of the partition

Conplete.SAT3 — Type

Instance of a 3-SAT problem.

Variables are denoted by integers from 1 to variable_count, negations of variables are denoted by corresponding negative integers.

Fields

variable_count : Number of different variables present in clauses

clauses : clauses denoted as matrix of size: 3 × m

Conplete.SAT3Solution — Type

Solution to a SAT3 problem

Fields

evaluation : BitVector denoting evaluation of each variable

Conplete.SubsetSum — Type

Instance of a Subset Sum problem

It utilizes generic integers, and can be constructed with any type which is a child of Integer abstract type

Fields

elements : vector of integers of which we seek a subset

sum : target sum

Conplete.SubsetSumSolution — Type

Solution to a SubsetSum problem

Fields

subset : generic set of indices corresponding to elements that are present in the subset

Conplete.TSP — Type

Instance of a Travelling Salesman problem

Fields

weights : weights of edges in a full undirected graph

length : maximal length of a cycle

Conplete.TSPSolution — Type

Solution to a TSP problem

Fields

cycle : Vector of length equal to number of vertices where value cycle[x] = y corresponds to edge (x,y) in cycle

Conplete.VertexCover — Type

Instance of a Vertex Cover problem.

Fields

graph : Undirected graph in which we seek a vertex cover

size : minimal size of a vertex cover

Conplete.VertexCoverSolution — Type

Solution to a VertexCover problem

Fields

clique : A generic set containing integers, corresponding to vertices belonging to vertex cover